Introduced new Analyser class structure, beginning with BTI analysis.
This is done to make the code in the main function clearer and to make the analysis of the given input samples
easier with multiple conditions.

Each analysis now should work by calling a wrapper class for each input file.
The Wrapper then calls the analysis classes for each subsystem. The plots can
be stored via calling savePlots on the wrapper.

The histograms, especially 2D, have to be made "colz"-Plots again.